+++ common/pom.xml
@@ -16,4 +16,13 @@
         <artifactId>neo4j-connector-apache-spark</artifactId>
         <version>4.0.0</version>
     </parent>
+
+    <dependencies>
+        <dependency>
+            <groupId>neo4j-contrib</groupId>
+            <artifactId>neo4j-connector-apache-spark_${scala.binary.version}_test-support</artifactId>
+            <version>4.0.0</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 </project>
