+++ common/src/test/scala/org/neo4j/spark/SparkConnectorScalaBaseWithApocTSE.scala
