+++ common/src/test/scala/org/neo4j/spark/SparkConnectorScalaSuiteIT.scala
@@ -8,14 +8,14 @@ import org.junit.{AfterClass, Assume, BeforeClass}
 import org.neo4j.Neo4jContainerExtension
 import org.neo4j.driver._
 import org.neo4j.driver.summary.ResultSummary
+import org.neo4j.spark.service.SchemaServiceTSE
 import org.neo4j.spark.util.Neo4jUtil
 
 
-object SparkConnectorScalaSuiteWithApocIT {
-  val server: Neo4jContainerExtension = new Neo4jContainerExtension()
+object SparkConnectorScalaSuiteIT {
+  val server: Neo4jContainerExtension = new Neo4jContainerExtension(s"neo4j:${TestUtil.neo4jVersion()}-enterprise")
     .withNeo4jConfig("dbms.security.auth_enabled", "false")
     .withEnv("NEO4J_ACCEPT_LICENSE_AGREEMENT", "yes")
-    .withEnv("NEO4JLABS_PLUGINS", "[\"apoc\"]")
     .withDatabases(Seq("db1", "db2"))
 
   var conf: SparkConf = _
@@ -32,7 +32,7 @@ object SparkConnectorScalaSuiteWithApocIT {
       try {
         server.start()
       } catch {
-        case _ => //
+        case _: Throwable => //
       }
       Assume.assumeTrue("Neo4j container is not started", server.isRunning)
       conf = new SparkConf().setAppName("neoTest")
@@ -83,7 +83,6 @@ object SparkConnectorScalaSuiteWithApocIT {
 
 @RunWith(classOf[Suite])
 @Suite.SuiteClasses(Array(
-  classOf[DataSourceReaderWithApocTSE],
-  classOf[DataSourceReaderNeo4j4xWithApocTSE]
+  classOf[SchemaServiceTSE],
 ))
-class SparkConnectorScalaSuiteWithApocIT {}
+class SparkConnectorScalaSuiteIT {}
