--- spark-2.4/src/test/resources/neo4j-spark-connector.properties
@@ -1,2 +1,0 @@
-neo4j.version=${neo4j.version}
-neo4j.experimental=${neo4j.experimental}
