+++ spark-3.0/src/main/scala/org/neo4j/spark/writer/Neo4jDataWriter.scala
@@ -1,130 +1,17 @@
 package org.neo4j.spark.writer
-import java.util
-import java.util.concurrent.CountDownLatch
-import org.apache.spark.internal.Logging
+
 import org.apache.spark.sql.SaveMode
 import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.connector.write.{DataWriter, WriterCommitMessage}
+import org.apache.spark.sql.connector.write.DataWriter
 import org.apache.spark.sql.types.StructType
-import org.neo4j.driver.exceptions.{ClientException, Neo4jException, ServiceUnavailableException, SessionExpiredException}
-import org.neo4j.driver.{Session, Transaction, Values}
-import org.neo4j.spark.service.{MappingService, Neo4jQueryService, Neo4jQueryStrategy, Neo4jQueryWriteStrategy, Neo4jWriteMappingStrategy}
-import org.neo4j.spark.util.{DriverCache, Neo4jOptions, Neo4jUtil}
-import org.neo4j.spark.util.Neo4jUtil.closeSafety
-
-import scala.collection.JavaConverters._
+import org.neo4j.spark.util.Neo4jOptions
 
 class Neo4jDataWriter(jobId: String,
                       partitionId: Int,
                       structType: StructType,
                       saveMode: SaveMode,
                       options: Neo4jOptions,
-                      scriptResult: java.util.List[java.util.Map[String, AnyRef]]) extends DataWriter[InternalRow] with Logging {
+                      scriptResult: java.util.List[java.util.Map[String, AnyRef]])
+  extends BaseDataWriter(jobId, partitionId, structType, saveMode, options, scriptResult)
+    with DataWriter[InternalRow]
 
-  private val driverCache: DriverCache = new DriverCache(options.connection, jobId)
-
-  private var transaction: Transaction = _
-  private var session: Session = _
-
-  private val mappingService = new MappingService(new Neo4jWriteMappingStrategy(options), options)
-
-  private val batch: util.List[java.util.Map[String, Object]] = new util.ArrayList[util.Map[String, Object]]()
-
-  private val retries = new CountDownLatch(options.transactionMetadata.retries)
-
-  val query: String = new Neo4jQueryService(options, new Neo4jQueryWriteStrategy(saveMode)).createQuery()
-
-  override def write(record: InternalRow): Unit = {
-    batch.add(mappingService.convert(record, structType))
-    if (batch.size() == options.transactionMetadata.batchSize) {
-      writeBatch()
-    }
-  }
-
-  private def writeBatch(): Unit = {
-    try {
-      if (session == null || !session.isOpen) {
-        session = driverCache.getOrCreate().session(options.session.toNeo4jSession)
-      }
-      if (transaction == null || !transaction.isOpen) {
-        transaction = session.beginTransaction()
-      }
-      log.info(
-        s"""Writing a batch of ${batch.size()} elements to Neo4j,
-           |for jobId=$jobId and partitionId=$partitionId
-           |with query: $query
-           |""".stripMargin)
-      val result = transaction.run(query,
-        Values.value(Map[String, AnyRef](Neo4jQueryStrategy.VARIABLE_EVENTS -> batch,
-          Neo4jQueryStrategy.VARIABLE_SCRIPT_RESULT -> scriptResult).asJava))
-      if (log.isDebugEnabled) {
-        val summary = result.consume()
-        val counters = summary.counters()
-        log.debug(
-          s"""Batch saved into Neo4j data with:
-             | - nodes created: ${counters.nodesCreated()}
-             | - nodes deleted: ${counters.nodesDeleted()}
-             | - relationships created: ${counters.relationshipsCreated()}
-             | - relationships deleted: ${counters.relationshipsDeleted()}
-             | - properties set: ${counters.propertiesSet()}
-             | - labels added: ${counters.labelsAdded()}
-             | - labels removed: ${counters.labelsRemoved()}
-             |""".stripMargin)
-      }
-      transaction.commit()
-      closeSafety(transaction)
-      batch.clear()
-    } catch {
-      case neo4jTransientException: Neo4jException => {
-        val code = neo4jTransientException.code()
-        if ((neo4jTransientException.isInstanceOf[SessionExpiredException] || neo4jTransientException.isInstanceOf[ServiceUnavailableException])
-          && !(Neo4jUtil.unsupportedTransientCodes ++ options.transactionMetadata.failOnTransactionCodes).contains(code)
-          && retries.getCount > 0) {
-          retries.countDown()
-          log.info(s"Matched Neo4j transient exception next retry is ${options.transactionMetadata.retries - retries.getCount}")
-          close()
-          writeBatch()
-        } else {
-          logAndThrowException(neo4jTransientException)
-        }
-      }
-      case e: Exception => logAndThrowException(e)
-    }
-    Unit
-  }
-
-  private def logAndThrowException(e: Exception): Unit = {
-    if (e.isInstanceOf[ClientException]) {
-      log.error(s"Cannot commit the transaction because: ${e.getMessage}")
-    }
-    else {
-      log.error("Cannot commit the transaction because the following exception", e)
-    }
-
-    throw e
-  }
-
-  override def commit(): WriterCommitMessage = {
-    writeBatch()
-    close()
-    null
-  }
-
-  override def abort(): Unit = {
-    if (transaction != null && transaction.isOpen) {
-      try {
-        transaction.rollback()
-      } catch {
-        case e: Throwable => log.warn("Cannot rollback the transaction because of the following exception", e)
-      }
-    }
-    close()
-    Unit
-  }
-
-
-  override def close(): Unit = {
-    closeSafety(transaction, log)
-    closeSafety(session, log)
-  }
-}
