+++ spark-3.0/src/test/scala/org/neo4j/spark/DataSourceReaderNeo4j4xWithApocTSE.scala
@@ -2,7 +2,6 @@ package org.neo4j.spark
 
 import org.junit.Assert.assertEquals
 import org.junit.{Assume, BeforeClass, Test}
-import org.neo4j.TestUtil
 import org.neo4j.driver.summary.ResultSummary
 import org.neo4j.driver.{SessionConfig, Transaction, TransactionWork}
